# .github/openapi.yml
openapi: "3.0.3"
info:
  title: TW Tournament App API
  version: "1.0.0"
servers:
  - url: http://localhost/api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
...
paths:
...
  /match/tournament/{tournamentId}:
    get:
      summary: Get all matches for a tournament
      operationId: getMatchesByTournament
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of matches
        "404":
          description: Tournament not found
  /match/{id}:
    get:
      summary: Get a match by ID
      operationId: getMatchById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match data
        "404":
          description: Not found
  /match:
    post:
      summary: Create a match (authenticated)
      operationId: createMatch
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tournamentId, player1, player2]
              properties:
                tournamentId:
                  type: string
                player1:
                  type: string
                player2:
                  type: string
      responses:
        "201":
          description: Match created
        "401":
          description: Not authenticated
  /match/{id}/report:
    patch:
      summary: Report match result (authenticated participant)
      operationId: reportResult
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winnerId]
              properties:
                winnerId:
                  type: string
      responses:
        "200":
          description: Result reported
        "401":
          description: Not authenticated
  /match/{id}/result:
    patch:
      summary: Record match result (authenticated)
      operationId: recordResult
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winnerId]
              properties:
                winnerId:
                  type: string
      responses:
        "200":
          description: Result recorded
        "401":
          description: Not authenticated
  /match/{id}/resolve:
    patch:
      summary: Resolve a disputed match (authenticated, creator)
      operationId: resolveDispute
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winnerId]
              properties:
                winnerId:
                  type: string
      responses:
        "200":
          description: Dispute resolved
        "401":
          description: Not authenticated
        "403":
          description: Not the tournament creator
  /match/{id}/override:
    patch:
      summary: Override match result (authenticated, creator)
      operationId: overrideResult
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winnerId]
              properties:
                winnerId:
                  type: string
      responses:
        "200":
          description: Result overridden
        "401":
          description: Not authenticated
  /match/{id}/status:
    patch:
      summary: Update match status (authenticated)
      operationId: updateMatchStatus
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, in_progress, completed, disputed]
      responses:
        "200":
          description: Status updated
        "401":
          description: Not authenticated

  # ── Stats ──────────────────────────────────────────────────────────────────
...